Added AuxiliaryObject, Other Updates
AuxiliaryObject is a subclass of BackendObject. Is meant to be used in conjunction with a User class or a FIPS class as an additional data point that itself would be saved to the backend. It is a multipurpose classes that is meant to be used as a Lock, Request Queue , and a Keychain.

AuxiliaryObjects contain a list of uids relalted to a singluar group of data points' uids.

- - -
Updates to Config file (SMALL)