Added User & FIPS Objects, Other Small Changes
User & FIPS classes are children to the BackendObject.

User is meant to represent an actual user of the application. A FIPS object represents a single county. Creating a user automatically creates a keychain for that user. Creating a FIPS object automatically creates a Lock & Request Queue for that county. It will also created a key for encryption.
- - - -
Updated some Menu Items