Scraping Logic
Will ignore entries that don't have a name.

Looks for book/page for all entries. If found, operates as it has. If not found, notes that in place of property address.
TODO: scan through this updated list, grab the ones that say "No Book/Page" and then do the convoluted search