Added GuerilloServer, removed GuerilloServer code from GuerilloClient/GUI
GuerilloServer is meant to be a standalone runnable for Guerillo
It is able to simply handle queries added to the backend and pull all of the data from the appropriate locations for them.
Currently can only handle a single query at a time and doesn't remove queries that another Server could potentially have handled after it picked it up.
Also need a UI interface for the Server so that you can visually see the work it's doing instead of needing to look directly at the Backend or at a GuerilloClient that requested data.
- - -
As it stands you can run the GuerilloClient directly from GuerilloClient.py
but not GuerilloServer from GuerilloServer.py (it must be run from __main__)
- - -
NOTE THIS HAS NOT YET BEEN FULLY TESTED FROM THE SAME COMPUTER OR DIFFERENT COMPUTERS AWAITING KEN'S EYES